move AdapterError into top level error.rs module
