+++ forc-plugins/forc-debug/src/error.rs
@@ -1,3 +1,5 @@
+use crate::types::Instruction;
+use dap::requests::Command;
 use thiserror::Error;
 
 #[derive(Debug, Error)]
@@ -34,6 +36,42 @@ pub enum ArgumentError {
 
     #[error("Invalid number format: {0}")]
     InvalidNumber(String),
+}
+
+#[derive(Error, Debug)]
+pub enum AdapterError {
+    #[error("Unhandled command")]
+    UnhandledCommand { command: Command },
+
+    #[error("Missing command")]
+    MissingCommand,
+
+    #[error("Missing configuration")]
+    MissingConfiguration,
+
+    #[error("Missing source path argument")]
+    MissingSourcePathArgument,
+
+    #[error("Missing breakpoint location")]
+    MissingBreakpointLocation,
+
+    #[error("Missing source map")]
+    MissingSourceMap { pc: Instruction },
+
+    #[error("Unknown breakpoint")]
+    UnknownBreakpoint { pc: Instruction },
+
+    #[error("Build failed")]
+    BuildFailed { reason: String },
+
+    #[error("No active test executor")]
+    NoActiveTestExecutor,
+
+    #[error("Test execution failed")]
+    TestExecutionFailed {
+        #[from]
+        source: anyhow::Error,
+    },
 }
 
 pub type Result<T> = std::result::Result<T, ForcDebugError>;
